% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_BASiCS.R
\name{BASiCS_method_definition}
\alias{BASiCS_method_definition}
\title{Get Information of BASiCS}
\usage{
BASiCS_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of BASiCS
}
\examples{
BASiCS_method_definition <- BASiCS_method_definition()

}
