% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17-BEARscc.R
\name{BEARscc_estimation}
\alias{BEARscc_estimation}
\title{Estimate Parameters From Real Datasets by BEARscc}
\usage{
BEARscc_estimation(ref_data, verbose = FALSE, other_prior = NULL, seed)
}
\arguments{
\item{ref_data}{A count matrix. Each row represents a gene and each column
represents a cell.}

\item{verbose}{Logical.}

\item{other_prior}{A list with names of certain parameters. Some methods need
extra parameters to execute the estimation step, so you must input them. When
you use BEARscc, you must input the \code{dilution.factor} and the \code{volume}
information to calculate the number of molecules of spike-ins.}

\item{seed}{An integer of a random seed.}
}
\value{
A list contains the estimated parameters and the results of execution
detection.
}
\description{
This function is used to estimate useful parameters from a real dataset by
using \code{estimate_noiseparameters} function in BEARscc package.
}
\references{
Severson D T, Owen R P, White M J, et al. BEARscc determines robustness of single-cell clusters using simulated technical replicates. Nature communications, 2018, 9(1): 1-7. \url{https://doi.org/10.1038/s41467-018-03608-y}

Bioconductor URL: \url{https://www.bioconductor.org/packages/release/bioc/html/BEARscc.html}

Github URL: \url{https://github.com/seversond12/BEARscc}
}
