% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_BEARscc.R
\name{BEARscc_method_definition}
\alias{BEARscc_method_definition}
\title{Get Information of BEARscc}
\usage{
BEARscc_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of BEARscc
}
\examples{
BEARscc_method_definition <- BEARscc_method_definition()

}
