% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_CancerInSilico.R
\name{CancerInSilico_method_definition}
\alias{CancerInSilico_method_definition}
\title{Get Information of CancerInSilico}
\usage{
CancerInSilico_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of CancerInSilico
}
\examples{
CancerInSilico_method_definition <- CancerInSilico_method_definition()
}
