% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_dropsim.R
\name{dropsim_method_definition}
\alias{dropsim_method_definition}
\title{Get Information of dropsim}
\usage{
dropsim_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of dropsim
}
\examples{
dropsim_method_definition <- dropsim_method_definition()

}
