% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24-dropsim.R
\name{dropsim_simulation}
\alias{dropsim_simulation}
\title{Simulate Datasets by dropsim}
\usage{
dropsim_simulation(
  parameters,
  other_prior = NULL,
  return_format,
  verbose = FALSE,
  seed
)
}
\arguments{
\item{parameters}{A object generated by \code{\link[dropsim:fit_parameters]{dropsim::fit_parameters()}}}

\item{other_prior}{A list with names of certain parameters. Some methods need
extra parameters to execute the estimation step, so you must input them. In
simulation step, the number of cells, genes, groups, batches, the percent of
DEGs are usually customed, so before simulating a dataset you must point it out.
See \code{Details} below for more information.}

\item{return_format}{A character. Alternative choices: list, SingleCellExperiment,
Seurat, h5ad. If you select \code{h5ad}, you will get a path where the .h5ad file saves to.}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
This function is used to simulate datasets from learned parameters by \code{simulateDGE}
function in dropsim package.
}
\details{
In dropsim, users can only set \code{nCells} and \code{nGenes} directly.
For more parameters in dropsim, check \code{\link[dropsim:simulateDGE]{dropsim::simulateDGE()}} and see \code{Examples}.
}
\examples{
\dontrun{
ref_data <- simmethods::data
## estimation
estimate_result <- simmethods::dropsim_estimation(
  ref_data = ref_data,
  verbose = TRUE,
  seed = 111
)

# 1) Simulate with default parameters
simulate_result <- simmethods::dropsim_simulation(
  parameters = estimate_result[["estimate_result"]],
  other_prior = NULL,
  return_format = "list",
  verbose = TRUE,
  seed = 111
)
## counts
counts <- simulate_result[["simulate_result"]][["count_data"]]
dim(counts)

# 2) 2000 cells and 5000 genes
simulate_result <- simmethods::dropsim_simulation(
  parameters = estimate_result[["estimate_result"]],
  other_prior = list(nCells = 2000,
                     nGenes = 5000),
  return_format = "list",
  verbose = TRUE,
  seed = 111
)

## counts
counts <- simulate_result[["simulate_result"]][["count_data"]]
dim(counts)
}

}
\references{
Github URL: \url{https://github.com/marchinilab/dropsim}
}
