% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_dyngen.R
\name{dyngen_method_definition}
\alias{dyngen_method_definition}
\title{Get Information of dyngen}
\usage{
dyngen_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of dyngen
}
\examples{
dyngen_method_definition <- dyngen_method_definition()

}
