% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_dyntoy.R
\name{dyntoy_method_definition}
\alias{dyntoy_method_definition}
\title{Get Information of dyntoy}
\usage{
dyntoy_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of dyntoy
}
\examples{
dyntoy_method_definition <- dyntoy_method_definition()
}
