% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_ESCO.R
\name{ESCO_method_definition}
\alias{ESCO_method_definition}
\title{Get Information of ESCO}
\usage{
ESCO_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of ESCO
}
\examples{
ESCO_method_definition <- ESCO_method_definition()

}
