% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20-ESCO.R
\name{ESCO_simulation}
\alias{ESCO_simulation}
\title{Simulate Datasets by ESCO}
\usage{
ESCO_simulation(
  parameters,
  return_format,
  other_prior = NULL,
  verbose = FALSE,
  seed
)
}
\arguments{
\item{parameters}{A object generated by \code{\link[ESCO:escoEstimate]{ESCO::escoEstimate()}}}

\item{return_format}{A character. Alternative choices: list, SingleCellExperiment,
Seurat, h5ad. If you select \code{h5ad}, you will get a path where the .h5ad file saves to.}

\item{other_prior}{A list with names of certain parameters. Some methods need
extra parameters to execute the estimation step, so you must input them. In
simulation step, the number of cells, genes, groups, batches, the percent of
DEGs are usually customed, so before simulating a dataset you must point it out.
See \code{Details} below for more information.}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
Simulate Datasets by ESCO
}
\references{
Tian J, Wang J, Roeder K. ESCO: single cell expression simulation incorporating gene co-expression. Bioinformatics, 2021, 37(16): 2374-2381. \url{https://doi.org/10.1093/bioinformatics/btab116}

Github URL: \url{https://github.com/JINJINT/ESCO}
}
