% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_manipulation.R
\name{get_method}
\alias{get_method}
\title{Get Information of Methods}
\usage{
get_method(method = "all")
}
\arguments{
\item{method}{One or more method names. Default is all.}
}
\value{
A list.
}
\description{
Get Information of Methods
}
\examples{
# get_method <- get_method(method = c("splat", "PROSSTT"))
}
