% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{group_condition}
\alias{group_condition}
\title{Group information of 160 Cells in GSE54695}
\format{
An object of class \code{factor} of length 160.
}
\usage{
group_condition
}
\description{
Group information of 160 Cells in GSE54695. Total 160 cells are
cultured in two types of medium, 0 for two-inhibitor (2i) and 1 for serum.
}
\keyword{datasets}
