% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/26-hierarchicell.R
\name{hierarchicell_estimation}
\alias{hierarchicell_estimation}
\title{Estimate Parameters From Real Datasets by hierarchicell}
\usage{
hierarchicell_estimation(ref_data, verbose = FALSE, other_prior = NULL, seed)
}
\arguments{
\item{ref_data}{A count matrix. Each row represents a gene and each column
represents a cell.}

\item{verbose}{Logical.}

\item{other_prior}{A list with names of certain parameters. Some methods need
extra parameters to execute the estimation step, so you must input them. In
simulation step, the number of cells, genes, groups, batches, the percent of
DEGs are usually customed, so before simulating a dataset you must point it out.
For more customed parameters, see \code{\link[hierarchicell:compute_data_summaries]{hierarchicell::compute_data_summaries()}}}

\item{seed}{An integer of a random seed.}
}
\value{
A list contains the estimated parameters and the results of execution
detection.
}
\description{
This function is used to estimate useful parameters from a real dataset by
using \code{compute_data_summaries} function in hierarchicell package.
}
\examples{
\dontrun{
ref_data <- SingleCellExperiment::counts(scater::mockSCE())
## estimation
estimate_result <- simmethods::hierarchicell_estimation(
  ref_data = ref_data,
  other_prior = NULL,
  verbose = TRUE,
  seed = 111
)
}

}
\references{
Zimmerman K D, Langefeld C D. Hierarchicell: an R-package for estimating power for tests of differential expression with single-cell data. BMC genomics, 2021, 22(1): 1-8. \url{https://doi.org/10.1186/s12864-021-07635-w}

Github URL: \url{https://github.com/kdzimm/hierarchicell}
}
