% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_hierarchicell.R
\name{hierarchicell_method_definition}
\alias{hierarchicell_method_definition}
\title{Get Information of hierarchicell}
\usage{
hierarchicell_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of hierarchicell
}
\examples{
hierarchicell_method_definition <- hierarchicell_method_definition()

}
