% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23-hierarchicell.R
\name{hierarchicell_simulation}
\alias{hierarchicell_simulation}
\title{Simulate Datasets by hierarchicell}
\usage{
hierarchicell_simulation(
  parameters,
  other_prior = NULL,
  return_format,
  verbose = FALSE,
  seed
)
}
\arguments{
\item{parameters}{A object generated by \code{\link[hierarchicell:compute_data_summaries]{hierarchicell::compute_data_summaries()}}}

\item{other_prior}{A list with names of certain parameters. Some methods need
extra parameters to execute the estimation step, so you must input them. In
simulation step, the number of cells, genes, groups, batches, the percent of
DEGs are usually customed, so before simulating a dataset you must point it out.
See \code{Details} below for more information.}

\item{return_format}{A character. Alternative choices: list, SingleCellExperiment,
Seurat, h5ad. If you select \code{h5ad}, you will get a path where the .h5ad file saves to.}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
This function is used to simulate datasets from learned parameters by \code{simulate_hierarchicell}
function in hierarchicell package.
}
\details{
In hierarchicell, users can set \code{nCells}, \code{nGenes} and \code{fc.group} directly.
There are some notes that users should know:
\enumerate{
\item hierarchicell can only simulate \strong{two} groups.
\item Some cells in the result may contain NA values across all genes due to the failing of GLM fitting.
\item hierarchicell does not return the information of DEGs and we can not know which genes are DEGs.
}

For more information, see \code{Examples} and \link[hierarchicell:simulate_hierarchicell]{hierarchicell::simulate_hierarchicell}
}
\examples{
\dontrun{
ref_data <- SingleCellExperiment::counts(scater::mockSCE())
## estimation
estimate_result <- simmethods::hierarchicell_estimation(
  ref_data = ref_data,
  other_prior = NULL,
  verbose = TRUE,
  seed = 111
)

# 1) Simulate with default parameters
simulate_result <- simmethods::hierarchicell_simulation(
  parameters = estimate_result[["estimate_result"]],
  other_prior = NULL,
  return_format = "list",
  verbose = TRUE,
  seed = 111
)
## counts
counts <- simulate_result[["simulate_result"]][["count_data"]]
dim(counts)

# 2) Customed parameters: cell and gene number, fold change of DEGs. (But hierarchicell
# does not tell us which genes are DEGs). Note that some cells may contain NA values
# across all genes due to the failing of GLM fitting.
simulate_result <- simmethods::hierarchicell_simulation(
  parameters = estimate_result[["estimate_result"]],
  other_prior = list(nCells = 2000,
                     nGenes = 4000,
                     fc.group = 4),
  return_format = "list",
  verbose = TRUE,
  seed = 111
)

## counts
counts <- simulate_result[["simulate_result"]][["count_data"]]
dim(counts)
## Remove NA cells
if(!requireNamespace("tidyr", quietly = TRUE)){utils::install.packages("tidyr")}
filter_counts <- as.matrix(t(tidyr::drop_na(as.data.frame(t(counts)))))
dim(filter_counts)
}

}
\references{
Zimmerman K D, Langefeld C D. Hierarchicell: an R-package for estimating power for tests of differential expression with single-cell data. BMC genomics, 2021, 22(1): 1-8. \url{https://doi.org/10.1186/s12864-021-07635-w}

Github URL: \url{https://github.com/kdzimm/hierarchicell}
}
