% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_Kersplat.R
\name{Kersplat_method_definition}
\alias{Kersplat_method_definition}
\title{Get Information of Kersplat}
\usage{
Kersplat_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of Kersplat
}
\examples{
Kersplat_method_definition <- Kersplat_method_definition()
}
