% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_Lun2.R
\name{Lun2_method_definition}
\alias{Lun2_method_definition}
\title{Get Information of Lun2}
\usage{
Lun2_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of Lun2
}
\examples{
Lun2_method_definition <- Lun2_method_definition()
}
