% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-Lun.R
\name{Lun_estimation}
\alias{Lun_estimation}
\title{Estimate Parameters From Real Datasets by Lun}
\usage{
Lun_estimation(ref_data, verbose = FALSE, seed)
}
\arguments{
\item{ref_data}{A count matrix. Each row represents a gene and each column
represents a cell.}

\item{verbose}{Logical.}

\item{seed}{An integer of a random seed.}
}
\value{
A list contains the estimated parameters and the results of execution
detection.
}
\description{
This function is used to estimate useful parameters from a real dataset by
using \code{lunEstimate} function in Splatter package.
}
