% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_Lun.R
\name{Lun_method_definition}
\alias{Lun_method_definition}
\title{Get Information of Lun}
\usage{
Lun_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of Lun
}
\examples{
Lun_method_definition <- Lun_method_definition()
}
