% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/34-MFA.R
\name{MFA_estimation}
\alias{MFA_estimation}
\title{Estimate Parameters From Real Datasets by MFA}
\usage{
MFA_estimation(ref_data, verbose = FALSE, seed)
}
\arguments{
\item{ref_data}{A count matrix. Each row represents a gene and each column
represents a cell.}

\item{verbose}{Logical.}

\item{seed}{An integer of a random seed.}
}
\value{
A list contains the estimated parameters and the results of execution
detection.
}
\description{
This function is used to estimate useful parameters from a real dataset by
using \code{splatEstimate} function in Splatter package.
}
\examples{
\dontrun{
ref_data <- simmethods::data
## estimation
estimate_result <- simmethods::MFA_estimation(
  ref_data = ref_data,
  verbose = TRUE,
  seed = 111
)
}

}
\references{
Campbell K R, Yau C. Probabilistic modeling of bifurcations in single-cell gene expression data using a Bayesian mixture of factor analyzers. Wellcome open research, 2017, 2. \url{https://doi.org/10.12688/wellcomeopenres.11087.1}

Github URL: \url{https://github.com/kieranrcampbell/mfa}
}
