% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_MFA.R
\name{MFA_method_definition}
\alias{MFA_method_definition}
\title{Get Information of MFA}
\usage{
MFA_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of MFA
}
\examples{
MFA_method_definition <- MFA_method_definition()
}
