% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32-MFA.R
\name{MFA_simulation}
\alias{MFA_simulation}
\title{Simulate Datasets by MFA}
\usage{
MFA_simulation(parameters, return_format, verbose = FALSE, seed)
}
\arguments{
\item{parameters}{A object generated by \code{\link[splatter]{mfaEstimate}}}

\item{return_format}{A character. Alternatives choices: list, SingleCellExperiment,
Seurat, h5ad}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
Simulate Datasets by MFA
}
