% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22-muscat.R
\name{muscat_estimation}
\alias{muscat_estimation}
\title{Estimate Parameters From Real Datasets by muscat}
\usage{
muscat_estimation(ref_data, verbose = FALSE, other_prior = NULL, seed)
}
\arguments{
\item{ref_data}{A count matrix. Each row represents a gene and each column
represents a cell.}

\item{verbose}{Logical.}

\item{other_prior}{A list with names of certain parameters. Some methods need
extra parameters to execute the estimation step, so you must input them. In
simulation step, the number of cells, genes, groups, batches, the percent of
DEGs and other variables are usually customed, so before simulating a dataset
you must point it out.}

\item{seed}{An integer of a random seed.}
}
\value{
A list contains the estimated parameters and the results of execution
detection.
}
\description{
This function is used to estimate useful parameters from a real dataset by
using \code{prepSim} function in muscat package.
}
\details{
In muscat, cell group information is not neccessary but users can input it when
it is available by \code{other_prior = list(group.condition = xxx)}.

For more information, see \code{Examples} and \code{\link[muscat:prepSim]{muscat::prepSim()}}
}
\examples{
\dontrun{
ref_data <- simmethods::data
## estimation
estimate_result <- simmethods::muscat_estimation(
  ref_data = ref_data,
  other_prior = NULL,
  verbose = TRUE,
  seed = 111
)

## cell groups
group_condition <- as.numeric(simmethods::group_condition)
## estimation
estimate_result <- simmethods::muscat_estimation(
  ref_data = ref_data,
  other_prior = list(group.condition = group_condition),
  verbose = TRUE,
  seed = 111
)
}

}
\references{
Crowell H L, Soneson C, Germain P L, et al. Muscat detects subpopulation-specific state transitions from multi-sample multi-condition single-cell transcriptomics data. Nature communications, 2020, 11(1): 1-12. \url{https://doi.org/10.1038/s41467-020-19894-4}

Github URL: \url{https://github.com/HelenaLC/muscat}
}
