% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_muscat.R
\name{muscat_method_definition}
\alias{muscat_method_definition}
\title{Get Information of muscat}
\usage{
muscat_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of muscat
}
\examples{
muscat_method_definition <- muscat_method_definition()

}
