% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_phenopath.R
\name{phenopath_method_definition}
\alias{phenopath_method_definition}
\title{Get Information of phenopath}
\usage{
phenopath_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of phenopath
}
\examples{
phenopath_method_definition <- phenopath_method_definition()

}
