% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-POWSC.R
\name{POWSC_estimation}
\alias{POWSC_estimation}
\title{Estimate Parameters From Real Datasets by POWSC}
\usage{
POWSC_estimation(ref_data, verbose = FALSE, other_prior, seed)
}
\arguments{
\item{ref_data}{A count matrix. Each row represents a gene and each column
represents a cell.}

\item{verbose}{Logical.}

\item{other_prior}{A list with names of certain parameters. Some methods need
extra parameters to execute the estimation step, so you must input them. In
simulation step, the number of cells, genes, groups, batches, the percent of
DEGs and other variables are usually customed, so before simulating a dataset
you must point it out.}

\item{seed}{An integer of a random seed.}
}
\value{
A list contains the estimated parameters and the results of execution
detection.
}
\description{
This function is used to estimate useful parameters from a real dataset by
using \code{Est2Phase} function in POWSC package.
}
\details{
When you use POWSC to estimate parameters from a real dataset, the default settings
are recommended.
}
\examples{
ref_data <- SingleCellExperiment::counts(scater::mockSCE())

## estimation
estimate_result <- simmethods::POWSC_estimation(ref_data = ref_data,
                                                other_prior = NULL,
                                                verbose = TRUE,
                                                seed = 111)
}
