% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_POWSC.R
\name{POWSC_method_definition}
\alias{POWSC_method_definition}
\title{Get Information of POWSC}
\usage{
POWSC_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of POWSC
}
\examples{
POWSC_method_definition <- POWSC_method_definition()

}
