% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-POWSC.R
\name{POWSC_simulation}
\alias{POWSC_simulation}
\title{Simulate Datasets by POWSC}
\usage{
POWSC_simulation(parameters, other_prior, return_format, verbose = FALSE, seed)
}
\arguments{
\item{parameters}{A object generated by \code{\link[POWSC:Est2Phase]{POWSC::Est2Phase()}}}

\item{other_prior}{A list with names of certain parameters. Some methods need
extra parameters to execute the estimation step, so you must input them. In
simulation step, the number of cells, genes, groups, batches, the percent of
DEGs and other variables are usually customed, so before simulating a dataset
you must point it out.}

\item{return_format}{A character. Alternative choices: list, SingleCellExperiment,
Seurat, h5ad. If you select \code{h5ad}, you will get a path where the .h5ad file saves to.}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
This function is used to simulate datasets from learned parameters by \code{Simulate2SCE}
function in POWSC package.
}
