% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_powsimR.R
\name{powsimR_method_definition}
\alias{powsimR_method_definition}
\title{Get Information of powsimR}
\usage{
powsimR_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of powsimR
}
\examples{
powsimR_method_definition <- powsimR_method_definition()

}
