% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_PROSSTT.R
\name{PROSSTT_method_definition}
\alias{PROSSTT_method_definition}
\title{Get Information of PROSSTT}
\usage{
PROSSTT_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of PROSSTT
}
\examples{
PROSSTT_method_definition <- PROSSTT_method_definition()
}
