% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/26-PROSSTT.R, R/27-TedSim.R
\name{PROSSTT_simulation}
\alias{PROSSTT_simulation}
\title{Simulate Datasets by PROSSTT}
\usage{
PROSSTT_simulation(parameters, return_format, verbose = FALSE, seed)

PROSSTT_simulation(parameters, return_format, verbose = FALSE, seed)
}
\arguments{
\item{parameters}{A object generated by \code{\link[simutils]{TedSim_est}}}

\item{return_format}{A character. Alternatives choices: list, SingleCellExperiment,
Seurat, h5ad}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}

\item{other_prior}{A list with names of certain parameters. Some methods need
extra parameters to execute the estimation step, so you must input them. In
simulation step, the number of cells, genes, groups, batches, the percent of
DEGs and other variables are usually customed, so before simulating a dataset
you must point it out.}
}
\description{
Simulate Datasets by PROSSTT

Simulate Datasets by PROSSTT
}
