% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_scDD.R
\name{scDD_method_definition}
\alias{scDD_method_definition}
\title{Get Information of scDD}
\usage{
scDD_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of scDD
}
\examples{
scDD_method_definition <- scDD_method_definition()

}
