% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_scDesign2.R
\name{scDesign2_method_definition}
\alias{scDesign2_method_definition}
\title{Get Information of scDesign2}
\usage{
scDesign2_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of scDesign2
}
\examples{
scDesign2_method_definition <- scDesign2_method_definition()

}
