% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-scDesign3.R
\name{scDesign3_estimation}
\alias{scDesign3_estimation}
\title{Estimate Parameters From Real Datasets by scDesign3}
\usage{
scDesign3_estimation(ref_data, other_prior = NULL, verbose = FALSE, seed)
}
\arguments{
\item{ref_data}{A count matrix. Each row represents a gene and each column
represents a cell.}

\item{other_prior}{A list with names of certain parameters. Some methods need
extra parameters to execute the estimation step, so you must input them. In
simulation step, the number of cells, genes, groups, batches, the percent of
DEGs are usually customed, so before simulating a dataset you must point it out.
See \code{Details} below for more information.}

\item{verbose}{Logical.}

\item{seed}{An integer of a random seed.}
}
\value{
A list contains the estimated parameters and the results of execution
detection.
}
\description{
This function is used to estimate useful parameters from a real dataset by
using \code{fit_marginal} and \code{fit_copula} function in scDesign3 package.
}
\details{
When you use scDesign3 to estimate parameters from a real dataset, you can optionally input
a numeric vector to specify the groups or plates that each cell comes from,
like \verb{other_prior = list(group.condition = the numeric vector)}.

You can also optionally input a numeric vector to specify the batches that each cell comes from,
like \verb{other_prior = list(batch.condition = the numeric vector)}.

See \code{Examples} and learn from it.
}
\examples{
\dontrun{
ref_data <- simmethods::data
estimate_result <- simmethods::Splat_estimation(ref_data = data,
                                                verbose = TRUE,
                                                seed = 10)
estimate_result <- estimate_result[["estimate_result"]]
## Check the class
class(estimate_result) == "SplatParams"
}

}
\references{
Song, D., Wang, Q., Yan, G. et al. scDesign3 generates realistic in silico data for multimodal single-cell and spatial omics. Nat Biotechnol (2023). https://doi.org/10.1038/s41587-023-01772-1

Github URL: \url{https://github.com/SONGDONGYUAN1994/scDesign3/tree/main}
}
