% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_scDesign3.R
\name{scDesign3_method_definition}
\alias{scDesign3_method_definition}
\title{Get Information of scDesign3}
\usage{
scDesign3_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of scDesign3
}
\examples{
muscat_method_definition <- muscat_method_definition()

}
