% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_scDesign.R
\name{scDesign_method_definition}
\alias{scDesign_method_definition}
\title{Get Information of scDesign}
\usage{
scDesign_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of scDesign
}
\examples{
scDesign_method_definition <- scDesign_method_definition()

}
