% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_scGAN.R
\name{scGAN_method_definition}
\alias{scGAN_method_definition}
\title{Get Information of scGAN}
\usage{
scGAN_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of scGAN
}
\examples{
scGAN_method_definition <- scGAN_method_definition()

}
