% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/36-scGAN.R
\name{scGAN_simulation}
\alias{scGAN_simulation}
\title{Simulate Datasets by scGAN}
\usage{
scGAN_simulation(parameters, return_format, verbose = FALSE, seed)
}
\arguments{
\item{parameters}{A object generated by \code{\link[=scGAN_estimation]{scGAN_estimation()}}}

\item{return_format}{A character. Alternatives choices: list, SingleCellExperiment,
Seurat, h5ad. If you select \code{h5ad}, you will get a path where the .h5ad file saves to.}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
This function is used to simulate datasets from learned parameters in the docker container.
}
\references{
Marouf M, Machart P, Bansal V, et al. Realistic in silico generation and augmentation of single-cell RNA-seq data using generative adversarial networks. Nature communications, 2020, 11(1): 1-12.
}
