% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_scMultiSim.R
\name{scMultiSim_method_definition}
\alias{scMultiSim_method_definition}
\title{Get Information of scMultiSim}
\usage{
scMultiSim_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of scMultiSim
}
\examples{
scMultiSim_method_definition <- scMultiSim_method_definition()

}
