% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31-scMultiSim.R
\name{scMultiSim_simulation}
\alias{scMultiSim_simulation}
\title{Simulate Datasets by scMultiSim}
\usage{
scMultiSim_simulation(
  parameters,
  other_prior = NULL,
  return_format,
  verbose = FALSE,
  seed
)
}
\arguments{
\item{parameters}{A object generated by \code{\link[simutils:make_trees]{simutils::make_trees()}}}

\item{other_prior}{A list with names of certain parameters. Some methods need
extra parameters to execute the estimation step, so you must input them. In
simulation step, the number of cells, genes, groups, batches, the percent of
DEGs are usually customed, so before simulating a dataset you must point it out.
See \code{Details} below for more information.}

\item{return_format}{A character. Alternative choices: list, SingleCellExperiment,
Seurat, h5ad. If you select \code{h5ad}, you will get a path where the .h5ad file saves to.}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
Simulate Datasets by scMultiSim
}
\details{
In scMultiSim, \code{nCells} and \code{nGenes} are usually customed and users can set
\code{other_prior = list(nCells = 1000, nGenes = 2000)} to simulate 1000 cells and 5000 genes.
In addition, \code{nBatches} can be customed for simulating the cell batches.
}
\examples{
ref_data <- simmethods::data
## estimation
estimate_result <- simmethods::scMultiSim_estimation(
  ref_data = ref_data,
  other_prior = NULL,
  verbose = TRUE,
  seed = 111
)

# 1) Simulate with default parameters
simulate_result <- simmethods::scMultiSim_simulation(
  parameters = estimate_result[["estimate_result"]],
  other_prior = NULL,
  return_format = "list",
  verbose = TRUE,
  seed = 111
)
## counts
counts <- simulate_result[["simulate_result"]][["count_data"]]
dim(counts)

}
\references{
Li H, Zhang Z, Squires M, et al. scMultiSim: simulation of multi-modality single cell data guided by cell-cell interactions and gene regulatory networks. bioRxiv, 2022: 2022.10. 15.512320. \url{https://doi.org/10.1101/2022.10.15.512320}

Github URL: \url{https://github.com/ZhangLabGT/scMultiSim}
}
