% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-SCRIP.R
\name{SCRIP_estimation}
\alias{SCRIP_estimation}
\title{Estimate Parameters From Real Datasets by SCRIP}
\usage{
SCRIP_estimation(ref_data, verbose = FALSE, seed)
}
\arguments{
\item{ref_data}{A count matrix. Each row represents a gene and each column
represents a cell.}

\item{verbose}{Logical.}

\item{seed}{An integer of a random seed.}
}
\value{
A list contains the estimated parameters and the results of execution
detection.
}
\description{
This function is used to estimate useful parameters from a real dataset by
using \code{splatEstimate} function in Splatter package.
}
\examples{
ref_data <- simmethods::data
estimate_result <- simmethods::SCRIP_estimation(ref_data = ref_data,
                                                verbose = TRUE,
                                                seed = 10)
estimate_result <- estimate_result[["estimate_result"]]
## Check the class (SCRIP accepts the estimation procedure in Splat)
class(estimate_result) == "SplatParams"
}
\references{
Qin F, Luo X, Xiao F, et al. SCRIP: an accurate simulator for single-cell RNA sequencing data\link{J}. Bioinformatics, 2022, 38(5): 1304-1311. \url{https://doi.org/10.1093/bioinformatics/btab824}

CRAN URL: \url{https://cran.r-project.org/web/packages/SCRIP/index.html}

Github URL: \url{https://github.com/thecailab/SCRIP}
}
