% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_SCRIP.R
\name{SCRIP_method_definition}
\alias{SCRIP_method_definition}
\title{Get Information of SCRIP}
\usage{
SCRIP_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of SCRIP
}
\examples{
Splat_method_definition <- Splat_method_definition()

}
