% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_SimBPDD.R
\name{SimBPDD_method_definition}
\alias{SimBPDD_method_definition}
\title{Get Information of SimBPDD}
\usage{
SimBPDD_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of SimBPDD
}
\examples{
muscat_method_definition <- muscat_method_definition()

}
