% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24-SimBPDD.R
\name{SimBPDD_simulation}
\alias{SimBPDD_simulation}
\title{Simulate Datasets by SimBPDD}
\usage{
SimBPDD_simulation(
  ref_data,
  other_prior = NULL,
  return_format,
  verbose = FALSE,
  seed
)
}
\arguments{
\item{ref_data}{A matrix for one dataset or a list of datasets with their own
names. This is usually unused except for some methods, e.g. SCRIP, SimBPDD,
zingeR.}

\item{other_prior}{A list with names of certain parameters. Some methods need
extra parameters to execute the estimation step, so you must input them. In
simulation step, the number of cells, genes, groups, batches, the percent of
DEGs are usually customed, so before simulating a dataset you must point it out.
See \code{Details} below for more information.}

\item{return_format}{A character. Alternatives choices: list, SingleCellExperiment,
Seurat, h5ad. If you select \code{h5ad}, you will get a path where the .h5ad file saves to.}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
This function is used to simulate datasets \code{design_data} function in SimBPDD package.
}
\details{
In addtion to simulate datasets with default parameters, users want to simulate
other kinds of datasets, e.g. a counts matrix with 2 or more cell groups. In
SimBPDD, you can set extra parameters to simulate datasets.

The customed parameters you can set are below:
\enumerate{
\item nCells. In SimBPDD, you can set nCells directly \code{other_prior = list(nCells = 1000)}
to simulate 1000 cells.
\item prob.group. You can directly set \code{other_prior = list(prob.group = c(0.4, 0.6))} to assign two proportions of cell groups. Note that the the length of the vector must be \strong{2}.
For more customed parameters in SimBPDD, please check \code{\link[SimBPDD:bp.sim.DD]{SimBPDD::bp.sim.DD()}}.
}
}
\references{
Schefzik R. SimBPDD: Simulating differential distributions in Beta-Poisson models, in particular for single-cell RNA sequencing data, Annales Mathematicae et Informaticae, 2021, 53: 283-298. \url{https://doi.org/10.33039/ami.2021.03.003}

Github URL: \url{https://github.com/RomanSchefzik/SimBPDD}
}
