% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmethods-package.R
\docType{package}
\name{simmethods-package}
\alias{simmethods}
\alias{simmethods-package}
\title{simmethods: A collection of 40+ simulation methods for singlce-cell RNA seqencing data}
\description{
A package contains the information and core execution functions of simulation methods for singlce-cell RNA seqencing data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/duohongrui/simmethods}
  \item Report bugs at \url{https://github.com/duohongrui/simmethods/issues}
}

}
\author{
\strong{Maintainer}: Hongrui Duo \email{Duohongrui@cqnu.edu.cn} (\href{https://orcid.org/0000-0001-8683-015X}{ORCID})

Authors:
\itemize{
  \item Bo Li \email{libo@cqnu.edu.cn} (\href{https://orcid.org/0000-0001-9944-0003}{ORCID})
}

}
\keyword{internal}
