% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_Simple.R
\name{Simple_method_definition}
\alias{Simple_method_definition}
\title{Get Information of Simple}
\usage{
Simple_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of Simple
}
\examples{
Simple_method_definition <- Simple_method_definition()
}
