% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_SparseDC.R
\name{SparseDC_method_definition}
\alias{SparseDC_method_definition}
\title{Get Information of SparseDC}
\usage{
SparseDC_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of SparseDC
}
\examples{
SparseDC_method_definition <- SparseDC_method_definition()

}
