% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_SPARSim.R
\name{SPARSim_method_definition}
\alias{SPARSim_method_definition}
\title{Get Information of SPARSim}
\usage{
SPARSim_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of SPARSim
}
\examples{
SPARSim_method_definition <- SPARSim_method_definition()

}
