% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18-SPARSim.R
\name{SPARSim_simulation}
\alias{SPARSim_simulation}
\title{Simulate Datasets by SPARSim}
\usage{
SPARSim_simulation(parameters, return_format, verbose = FALSE, seed)
}
\arguments{
\item{parameters}{A object generated by \code{\link[SPARSim]{SPARSim_simulation}}}

\item{return_format}{A character. Alternatives choices: list, SingleCellExperiment,
Seurat, h5ad}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
Simulate Datasets by SPARSim
}
