% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_splat.R
\name{splat_method_definition}
\alias{splat_method_definition}
\title{Get Information of Splat}
\usage{
splat_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of Splat
}
\examples{
splat_method_definition <- splat_method_definition()
}
