% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_SplatPop.R
\name{SplatPop_method_definition}
\alias{SplatPop_method_definition}
\title{Get Information of SplatPop}
\usage{
SplatPop_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of SplatPop
}
\examples{
SplatPop_method_definition <- SplatPop_method_definition()
}
