% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_SPsimSeq.R
\name{SPsimSeq_method_definition}
\alias{SPsimSeq_method_definition}
\title{Get Information of SPsimSeq}
\usage{
SPsimSeq_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of SPsimSeq
}
\examples{
SPsimSeq_method_definition <- SPsimSeq_method_definition()

}
