% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_SRTsim.R
\name{SRTsim_method_definition}
\alias{SRTsim_method_definition}
\title{Get Information of SRTsim}
\usage{
SRTsim_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of SRTsim
}
\examples{
SRTsim_method_definition <- SRTsim_method_definition()

}
