% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/28-SRTsim.R
\name{SRTsim_simulation}
\alias{SRTsim_simulation}
\title{Simulate Datasets by SRTsim}
\usage{
SRTsim_simulation(
  parameters,
  other_prior = NULL,
  return_format,
  verbose = FALSE,
  seed
)
}
\arguments{
\item{parameters}{A object generated by \code{\link[SRTsim:srtsim_fit]{SRTsim::srtsim_fit()}}}

\item{other_prior}{A list with names of certain parameters in SRTsim.}

\item{return_format}{A character. Alternative choices: list, SingleCellExperiment,
Seurat, h5ad. If you select \code{h5ad}, you will get a path where the .h5ad file saves to.}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
This function is used to simulate datasets from learned parameters by \code{srtsim_count}
function in SRTsim package.
}
\details{
For more customed parameters in SRTsim, please check \code{\link[SRTsim:srtsim_fit]{SRTsim::srtsim_fit()}}.
For detailed information about SRTsim, go to \url{https://cran.r-project.org/web/packages/SRTsim/index.html}.
}
\references{
Zhu J, Shang L, Zhou X. SRTsim: spatial pattern preserving simulations for spatially resolved transcriptomics. Genome Biology, 2023, 24(1): 39. \url{https://doi.org/10.1186/s13059-023-02879-z}

URL: \url{https://cran.r-project.org/web/packages/SRTsim/index.html}
}
