% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_SymSim.R
\name{SymSim_method_definition}
\alias{SymSim_method_definition}
\title{Get Information of SymSim}
\usage{
SymSim_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of SymSim
}
\examples{
SymSim_method_definition <- SymSim_method_definition()
}
