% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/27-TedSim.R
\name{TedSim_est}
\alias{TedSim_est}
\title{Bundled TedSim Estimation Function}
\usage{
TedSim_est(ref_data, phyla, seed)
}
\arguments{
\item{ref_data}{Reference dataset}
}
\value{
A list
}
\description{
Bundled TedSim Estimation Function
}
