% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/27-TedSim.R
\name{TedSim_estimation}
\alias{TedSim_estimation}
\title{Estimate Parameters From Real Datasets by TedSim}
\usage{
TedSim_estimation(ref_data, other_prior = NULL, verbose = FALSE, seed)
}
\arguments{
\item{ref_data}{A count matrix. Each row represents a gene and each column
represents a cell.}

\item{other_prior}{A list with names of certain parameters. Some methods need
extra parameters to execute the estimation step, so you must input them. In
simulation step, the number of cells, genes, groups, batches, the percent of
DEGs are usually customed, so before simulating a dataset you must point it out.
See \code{Details} below for more information.}

\item{verbose}{Logical.}

\item{seed}{An integer of a random seed.}
}
\value{
A list contains the estimated parameters and the results of execution
detection.
}
\description{
This function is used to estimate useful parameters from a real dataset by
using \code{TedSim_est} function in simmethods package.
}
\details{
In TedSim, users can input cell group information if it is available. If cell
group information is not provided, the procedure will detect cell groups by
kmeans automatically.
See \code{Examples} for more instructions.

Note that TedSim can only simulate the dataset whose cell number is the power of 2
so if the reference data does not meet the requirement, the procedure will
synthesize extra fake cells to achive this goal. See \code{\link[simutils:synthesize_cells]{simutils::synthesize_cells()}}
to learn about the process of synthesizing fake cells.
}
\examples{
ref_data <- simmethods::data
## estimation
estimate_result <- simmethods::TedSim_estimation(
  ref_data = ref_data,
  other_prior = NULL,
  verbose = TRUE,
  seed = 111
)

## estimation with cell group information
group_condition <- paste0("Group", as.numeric(simmethods::group_condition))
estimate_result <- simmethods::TedSim_estimation(
  ref_data = ref_data,
  other_prior = list(group.condition = group_condition),
  verbose = TRUE,
  seed = 111
)
}
\references{
Pan X, Li H, Zhang X. TedSim: temporal dynamics simulation of single-cell RNA sequencing data and cell division history. Nucleic acids research, 2022, 50(8): 4272-4288. \url{https://doi.org/10.1093/nar/gkac235}

Github URL: \url{https://github.com/Galaxeee/TedSim}
}
