% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_TedSim.R
\name{TedSim_method_definition}
\alias{TedSim_method_definition}
\title{Get Information of TedSim}
\usage{
TedSim_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of TedSim
}
\examples{
TedSim_method_definition <- TedSim_method_definition()
}
