% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_VeloSim.R
\name{VeloSim_method_definition}
\alias{VeloSim_method_definition}
\title{Get Information of VeloSim}
\usage{
VeloSim_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of VeloSim
}
\examples{
VeloSim_method_definition <- VeloSim_method_definition()
}
