% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/35-VeloSim.R
\name{VeloSim_simulation}
\alias{VeloSim_simulation}
\title{Simulate Datasets by VeloSim}
\usage{
VeloSim_simulation(
  parameters,
  other_prior = NULL,
  return_format,
  verbose = FALSE,
  seed
)
}
\arguments{
\item{parameters}{A object generated by \code{\link[simutils:make_trees]{simutils::make_trees()}}}

\item{other_prior}{A list with names of certain parameters. Some methods need
extra parameters to execute the estimation step, so you must input them. In
simulation step, the number of cells, genes, groups, batches, the percent of
DEGs and other variables are usually customed, so before simulating a dataset
you must point it out.}

\item{return_format}{A character. Alternatives choices: list, SingleCellExperiment,
Seurat, h5ad}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
Simulate Datasets by VeloSim
}
\details{
In VeloSim, users can only set \code{nCells} and \code{nGenes} to specify the number of cells and genes in the
simulated dataset. See \code{Examples} for instructions.
}
\examples{
\dontrun{
ref_data <- simmethods::data

## estimation with cell group information
group_condition <- paste0("Group", as.numeric(simmethods::group_condition))
estimate_result <- simmethods::VeloSim_estimation(
  ref_data = ref_data,
  other_prior = list(group.condition = group_condition),
  verbose = TRUE,
  seed = 111
)

# 1) Simulate with default parameters
simulate_result <- simmethods::VeloSim_simulation(
  parameters = estimate_result[["estimate_result"]],
  other_prior = NULL,
  return_format = "list",
  verbose = TRUE,
  seed = 111
)
## counts
counts <- simulate_result[["simulate_result"]][["count_data"]]
dim(counts)

# 2) 200 cells and 5000 genes
simulate_result <- simmethods::VeloSim_simulation(
  parameters = estimate_result[["estimate_result"]],
  other_prior = list(nCells = 200,
                     nGenes = 5000),
  return_format = "list",
  verbose = TRUE,
  seed = 111
)

## counts
counts <- simulate_result[["simulate_result"]][["count_data"]]
dim(counts)
}

}
\references{
Zhang Z, Zhang X. VeloSim: Simulating single cell gene-expression and RNA velocity. BioRxiv, 2021. \url{https://doi.org/10.1101/2021.01.11.426277}

Github URL: \url{https://github.com/PeterZZQ/VeloSim}
}
