% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_zinbwave.R
\name{zinbwave_method_definition}
\alias{zinbwave_method_definition}
\title{Get Information of zinbwave}
\usage{
zinbwave_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of zinbwave
}
\examples{
zinbwave_method_definition <- zinbwave_method_definition()
}
