% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-zinbwave.R
\name{zinbwave_simulation}
\alias{zinbwave_simulation}
\title{Simulate Datasets by zinbwave}
\usage{
zinbwave_simulation(parameters, return_format, verbose = FALSE, seed)
}
\arguments{
\item{parameters}{A object generated by \code{\link[splatter]{zinbEstimate}}}

\item{return_format}{A character. Alternatives choices: list, SingleCellExperiment,
Seurat, h5ad}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
Simulate Datasets by zinbwave
}
