% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_zinbwaveZinger.R
\name{zinbwaveZinger_method_definition}
\alias{zinbwaveZinger_method_definition}
\title{Get Information of zinbwaveZinger}
\usage{
zinbwaveZinger_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of zinbwaveZinger
}
\examples{
zinbwaveZinger_method_definition <- zinbwaveZinger_method_definition()
}
