% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_zingeR.R
\name{zingeR_method_definition}
\alias{zingeR_method_definition}
\title{Get Information of zingeR}
\usage{
zingeR_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of zingeR
}
\examples{
zingeR_method_definition <- zingeR_method_definition()
}
