% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20-ESCO.R
\name{ESCO_simulation}
\alias{ESCO_simulation}
\title{Simulate Datasets by ESCO}
\usage{
ESCO_simulation(parameters, return_format, verbose = FALSE, seed)
}
\arguments{
\item{parameters}{A object generated by \code{\link[ESCO]{escoEstimate}}}

\item{return_format}{A character. Alternatives choices: list, SingleCellExperiment,
Seurat, h5ad}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
Simulate Datasets by ESCO
}
