% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23-hierarchicell.R
\name{hierarchicell_simulation}
\alias{hierarchicell_simulation}
\title{Simulate Datasets by hierarchicell}
\usage{
hierarchicell_simulation(
  parameters,
  other_prior = NULL,
  return_format,
  verbose = FALSE,
  seed
)
}
\arguments{
\item{parameters}{A object generated by \code{\link[hierarchicell:simulate_hierarchicell]{hierarchicell::simulate_hierarchicell()}}}

\item{other_prior}{A list with names of certain parameters. Some methods need
extra parameters to execute the estimation step, so you must input them. In
simulation step, the number of cells, genes, groups, batches, the percent of
DEGs are usually customed, so before simulating a dataset you must point it out.
See \code{Details} below for more information.}

\item{return_format}{A character. Alternative choices: list, SingleCellExperiment,
Seurat, h5ad. If you select \code{h5ad}, you will get a path where the .h5ad file saves to.}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
Simulate Datasets by hierarchicell
}
\references{
Zimmerman K D, Langefeld C D. Hierarchicell: an R-package for estimating power for tests of differential expression with single-cell data. BMC genomics, 2021, 22(1): 1-8. \url{https://doi.org/10.1186/s12864-021-07635-w}

Github URL: \url{https://github.com/kdzimm/hierarchicell}
}
