% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23-hierarchicell.R
\name{hierarchicell_simulation}
\alias{hierarchicell_simulation}
\title{Simulate Datasets by hierarchicell}
\usage{
hierarchicell_simulation(
  parameters,
  other_prior,
  return_format,
  verbose = FALSE,
  seed
)
}
\arguments{
\item{parameters}{A object generated by \code{\link[hierarchicell]{simulate_hierarchicell}}}

\item{other_prior}{A list with names of certain parameters. Some methods need
extra parameters to execute the estimation step, so you must input them. In
simulation step, the number of cells, genes, groups, batches, the percent of
DEGs and other variables are usually customed, so before simulating a dataset
you must point it out.}

\item{return_format}{A character. Alternatives choices: list, SingleCellExperiment,
Seurat, h5ad}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
Simulate Datasets by hierarchicell
}
