% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-Kersplat.R
\name{Kersplat_estimation}
\alias{Kersplat_estimation}
\title{Estimate Parameters From Real Datasets by Kersplat}
\usage{
Kersplat_estimation(ref_data, verbose = FALSE, seed)
}
\arguments{
\item{ref_data}{A count matrix. Each row represents a gene and each column
represents a cell.}

\item{verbose}{Logical.}

\item{seed}{An integer of a random seed.}
}
\value{
A list contains the estimated parameters and the results of execution
detection.
}
\description{
This function is used to estimate useful parameters from a real dataset by
using \code{kersplatEstimate} function in Splatter package.
}
\examples{
ref_data <- simmethods::data
# Estimate parameters
estimate_result <- Kersplat_estimation(ref_data = ref_data,
                                       verbose = TRUE,
                                       seed = 111)
estimate_result <- estimate_result[["estimate_result"]]
## Check the class
class(estimate_result) == "KersplatParams"
}
\references{
Zappia, L., Phipson, B. & Oshlack, A. Splatter: simulation of single-cell RNA sequencing data. Genome Biology 18, 174 (2017). \url{https://doi.org/10.1186/s13059-017-1305-0}

Bioconductor URL: \url{https://bioconductor.org/packages/release/bioc/html/splatter.html}

Github URL: \url{https://github.com/Oshlack/splatter}
}
