% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-Lun.R
\name{Lun_estimation}
\alias{Lun_estimation}
\title{Estimate Parameters From Real Datasets by Lun}
\usage{
Lun_estimation(ref_data, verbose = FALSE, seed)
}
\arguments{
\item{ref_data}{A count matrix. Each row represents a gene and each column
represents a cell.}

\item{verbose}{Logical.}

\item{seed}{An integer of a random seed.}
}
\value{
A list contains the estimated parameters and the results of execution
detection.
}
\description{
This function is used to estimate useful parameters from a real dataset by
using \code{lunEstimate} function in Splatter package.
}
\examples{
ref_data <- simmethods::data
estimate_result <- simmethods::Lun_estimation(ref_data = data,
                                              verbose = TRUE,
                                              seed = 10)
estimate_result <- estimate_result[["estimate_result"]]
## Check the class
class(estimate_result) == "LunParams"
}
\references{
Zappia L, Phipson B, Oshlack A. Splatter: simulation of single-cell RNA sequencing data. Genome biology, 2017, 18(1): 1-15. https://doi.org/10.1186/s13059-017-1305-0

Bioconductor URL: \url{https://bioconductor.org/packages/release/bioc/html/splatter.html}

Github URL: \url{https://github.com/Oshlack/splatter}
}
