% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-POWSC.R
\name{POWSC_simulation}
\alias{POWSC_simulation}
\title{Simulate Datasets by POWSC}
\usage{
POWSC_simulation(
  parameters,
  other_prior = NULL,
  return_format,
  verbose = FALSE,
  seed
)
}
\arguments{
\item{parameters}{A object generated by \code{\link[POWSC:Est2Phase]{POWSC::Est2Phase()}}}

\item{other_prior}{A list with names of certain parameters. Some methods need
extra parameters to execute the estimation step, so you must input them. In
simulation step, the number of cells, genes, groups, batches, the percent of
DEGs and other variables are usually customed, so before simulating a dataset
you must point it out.}

\item{return_format}{A character. Alternative choices: list, SingleCellExperiment,
Seurat, h5ad. If you select \code{h5ad}, you will get a path where the .h5ad file saves to.}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
This function is used to simulate datasets from learned parameters by \code{Simulate2SCE}
function in POWSC package.
}
\details{
In addtion to simulate datasets with default parameters, users want to simulate
other kinds of datasets, e.g. a counts matrix with 2 or more cell groups. In
POWSC, you can set extra parameters to simulate datasets.

The customed parameters you can set are below:
\enumerate{
\item nCells. In POWSC, you can set nCells directly by \code{other_prior = list(nCells = 1000)}.
\item nGroups. POWSC can only simulate \strong{two} groups.
\item de.prob. You can directly set \code{other_prior = list(de.prob = 0.2)} to simulate DEGs that account for 20 percent of all genes.
}
}
\examples{
\dontrun{
ref_data <- SingleCellExperiment::counts(scater::mockSCE())

## estimation
estimate_result <- simmethods::POWSC_estimation(ref_data = ref_data,
                                                other_prior = NULL,
                                                verbose = TRUE,
                                                seed = 111)

## default setting
simulate_result <- simmethods::POWSC_simulation(
  parameters = estimate_result[["estimate_result"]],
  other_prior = NULL,
  return_format = "list",
  verbose = TRUE,
  seed = 111
)
## counts
counts <- simulate_result[["simulate_result"]][["count_data"]]
dim(counts)
## cell information
col_data <- simulate_result[["simulate_result"]][["col_meta"]]
table(col_data$group)
## gene information
row_data <- simulate_result[["simulate_result"]][["row_meta"]]
table(row_data$de_gene)


## Simulate 1000 cells (de.prob = 0.2)
simulate_result <- simmethods::POWSC_simulation(
  parameters = estimate_result[["estimate_result"]],
  other_prior = list(nCells = 1000,
                     de.prob = 0.2),
  return_format = "list",
  verbose = TRUE,
  seed = 111
)
## counts
counts <- simulate_result[["simulate_result"]][["count_data"]]
dim(counts)
## cell information
col_data <- simulate_result[["simulate_result"]][["col_meta"]]
table(col_data$group)
## gene information
row_data <- simulate_result[["simulate_result"]][["row_meta"]]
table(row_data$de_gene)
}

}
