% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_Splat.R
\name{Splat_method_definition}
\alias{Splat_method_definition}
\title{Get Information of Splat}
\usage{
Splat_method_definition(...)
}
\arguments{
\item{...}{...}
}
\value{
A list contains the information of method and default parameters
}
\description{
Get Information of Splat
}
\examples{
Splat_method_definition <- Splat_method_definition()
}
