% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19-SPsimSeq.R
\name{SPsimSeq_simulation}
\alias{SPsimSeq_simulation}
\title{Simulate Datasets by SPsimSeq}
\usage{
SPsimSeq_simulation(ref_data, return_format, verbose = FALSE, seed)
}
\arguments{
\item{ref_data}{A matrix for one dataset or a list of datasets with their own
names. This is usually unused except for some methods, e.g. SCRIP, scDesign,
zingeR, SPsimSeq}

\item{return_format}{A character. Alternatives choices: list, SingleCellExperiment,
Seurat, h5ad}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
Simulate Datasets by SPsimSeq
}
