% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/27-TedSim.R
\name{TedSim_simulation}
\alias{TedSim_simulation}
\title{Simulate Datasets by TedSim}
\usage{
TedSim_simulation(parameters, return_format, verbose = FALSE, seed)
}
\arguments{
\item{parameters}{A object generated by \code{\link[simmethods]{TedSim_est}}}

\item{return_format}{A character. Alternatives choices: list, SingleCellExperiment,
Seurat, h5ad}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
Simulate Datasets by TedSim
}
