% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/27-TedSim.R
\name{TedSim_simulation}
\alias{TedSim_simulation}
\title{Simulate Datasets by TedSim}
\usage{
TedSim_simulation(parameters, return_format, verbose = FALSE, seed)
}
\arguments{
\item{parameters}{A object generated by \code{\link[=TedSim_est]{TedSim_est()}}}

\item{return_format}{A character. Alternative choices: list, SingleCellExperiment,
Seurat, h5ad. If you select \code{h5ad}, you will get a path where the .h5ad file saves to.}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
Simulate Datasets by TedSim
}
\references{
Pan X, Li H, Zhang X. TedSim: temporal dynamics simulation of single-cell RNA sequencing data and cell division history. Nucleic acids research, 2022, 50(8): 4272-4288. \url{https://doi.org/10.1093/nar/gkac235}

Github URL: \url{https://github.com/Galaxeee/TedSim}
}
