% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-zinbwave.R
\name{zinbwave_simulation}
\alias{zinbwave_simulation}
\title{Simulate Datasets by zinbwave}
\usage{
zinbwave_simulation(parameters, return_format, verbose = FALSE, seed)
}
\arguments{
\item{parameters}{A object generated by \code{\link[splatter:zinbEstimate]{splatter::zinbEstimate()}}}

\item{return_format}{A character. Alternative choices: list, SingleCellExperiment,
Seurat, h5ad. If you select \code{h5ad}, you will get a path where the .h5ad file saves to.}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
Simulate Datasets by zinbwave
}
\examples{
ref_data <- simmethods::data

estimate_result <- simmethods::zinbwave_estimation(ref_data = ref_data,
                                                   other_prior = NULL,
                                                   verbose = TRUE,
                                                   seed = 111)

## In zinbwave, users can not set the number of cells and genes.
simulate_result <- simmethods::zinbwave_simulation(
  parameters = estimate_result[["estimate_result"]],
  return_format = "list",
  verbose = TRUE,
  seed = 111
)
}
\references{
Zappia L, Phipson B, Oshlack A. Splatter: simulation of single-cell RNA sequencing data. Genome biology, 2017, 18(1): 1-15. \url{https://doi.org/10.1186/s13059-017-1305-0}

Bioconductor URL: \url{https://bioconductor.org/packages/release/bioc/html/splatter.html}

Github URL: \url{https://github.com/Oshlack/splatter}
}
