% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-zinbwaveZinger.R
\name{zinbwaveZinger_simulation}
\alias{zinbwaveZinger_simulation}
\title{Simulate Datasets by zingeR}
\usage{
zinbwaveZinger_simulation(
  ref_data,
  parameters,
  return_format,
  verbose = FALSE,
  seed
)
}
\arguments{
\item{ref_data}{A matrix for one dataset or a list of datasets with their own
names. This is usually unused except for some methods, e.g. SCRIP, scDesign,
zingeR, zinbwaveZinger.}

\item{parameters}{A object generated by \code{\link[zinbwaveZingercollected]{NBsimSingleCell_zinbwaveZinger}}}

\item{return_format}{A character. Alternatives choices: list, SingleCellExperiment,
Seurat, h5ad}

\item{verbose}{Logical. Whether to return messages or not.}

\item{seed}{A random seed.}
}
\description{
Simulate Datasets by zingeR
}
